scraper_service/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
httpx
trafilatura
pinecone>=3.0.0
langchain-text-splitters
PyJWT
